@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <Version>0.9.0.537</Version>
     <Configurations>Debug;Release;DebugLocal</Configurations>
     <ProjectGuid>8544FAAC-CC7F-4330-9C19-61A6FC91182C</ProjectGuid>
